+++ src/core/injections/semiblind/techniques/tempfile_based/tfb_handler.py
@@ -281,8 +281,8 @@ def tfb_injection_handler(url, timesec, filename, tmp_path, http_request_method,
                       settings.FOUND_DIFF = how_long - timesec
                       if false_positive_warning:
                         time.sleep(1)
-                      randv1 = random.randrange(0, 1)
-                      randv2 = random.randrange(1, 2)
+                      randv1 = random.randrange(0, 4)
+                      randv2 = random.randrange(1, 5)
                       randvcalc = randv1 + randv2
 
                       if settings.TARGET_OS == "win":